fix: Use new WorldsListPage
