+++ src/components/Modals/WorldsYourStorageModal/WorldsYourStorageModal.tsx
@@ -4,7 +4,7 @@ import Modal from 'decentraland-dapps/dist/containers/Modal'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { formatNumber } from 'decentraland-dapps/dist/lib/utils'
 import { track } from 'modules/analytics/sagas'
-import { fromBytesToMegabytes } from 'components/WorldListPage_WorldsForEnsOwnersFeature/utils'
+import { fromBytesToMegabytes } from 'components/WorldListPage/utils'
 import { config } from 'config'
 import { InfoIcon } from 'components/InfoIcon'
 import goodImg from './images/good.svg'
