--- src/components/WorldListPage_WorldsForEnsOwnersFeature/index.ts
@@ -1,2 +1,0 @@
-import WorldListPage from './WorldListPage.container'
-export default WorldListPage
