+++ src/routing/Routes.types.ts
@@ -2,7 +2,6 @@ import { RouteComponentProps } from 'react-router'
 
 export type Props = RouteComponentProps & {
   inMaintenance: boolean
-  isWorldsForEnsOwnersEnabled: boolean
 }
 
 export type MapStateProps = Pick<Props, 'inMaintenance'>
